fix(promotion): auto-promotion fails closed on unscanned artifacts
PromotionRuleService::evaluate_artifact skipped the CVE severity check when
no completed scan existed, so try_auto_promote auto-promoted unscanned
artifacts even when a max_cve_severity rule required scanning. PR #1648
fixed only the manual PromotionPolicyService path; this closes the parallel
auto-promotion path.

Extract the scan-state classifier (ScanState/ScanStateRow/classify_scan_state)
into a shared services::scan_state module so both promotion paths use one
table-tested definition of "unscanned". The rule path now classifies scan
state when no completed scan exists and records a violation when the artifact
is genuinely unscanned (never-scanned, in-progress, or crashed scanner),
mirroring evaluate_block_unscanned. not_applicable and completed are never
treated as unscanned, so we do not over-block.
